+++ libvpvl/src/FaceKeyFrame.cc
@@ -76,7 +76,7 @@ void FaceKeyFrame::write(uint8_t *data)
 {
     FaceKeyFrameChunk chunk;
     copyBytesSafe(chunk.name, m_name, sizeof(chunk.name));
-    chunk.frameIndex = m_frameIndex;
+    chunk.frameIndex = static_cast<uint32_t>(m_frameIndex);
     chunk.weight = m_weight;
     internal::copyBytes(data, reinterpret_cast<const uint8_t *>(&chunk), sizeof(chunk));
 }
